backend/requirements.txt
asgiref==3.5.2
boto3==1.24.28
botocore==1.27.28
dj-database-url==0.5.0
Django==4.0.5
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.1.1
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
